+++ R/feedback.R
@@ -55,7 +55,7 @@
 #' @export
 feedback <- function(fit, quantiles =  NA, values = NA, dist= "best", ex = NA, sf = 3){
   if(nrow(fit$vals)>1 & is.na(ex)==T){
-    return(feedbackgroup(fit, quantiles, values, dist, sf))
+    return(feedbackgroup(fit, quantiles, values, dist, sfg = sf))
   }
   
   if(nrow(fit$vals)>1 & is.na(ex)==F){