+++ R/condDirichlet.R
@@ -1,7 +1,0 @@
-#' @export
-condDirichlet <- function(){
-  cat("condDirichlet() has been removed\n")
-  cat("use elicitDirichlet() instead for eliciting a Dirichlet distribution\n")
-  cat("and exploring conditional distribuitions.\n")
-}
-