+++ R/deprecatedFunctions.R
@@ -1,0 +1,32 @@
+#' @export
+
+roulette <- function(){
+  cat("roulette() has been removed\n")
+  cat("the roulette method is now available within elicit()\n")
+  
+}
+
+#' @export
+elicitQuartiles <- function(){
+  cat("elicitQuartiles() has been removed\n")
+  cat("quartiles elicitation can be performed using elicit()\n")  
+}
+
+#' @export
+elicitTertiles <- function(){
+  cat("elicitTertiles() has been removed\n")
+  cat("tertiles elicitation can be performed using elicit()\n")  
+}
+
+#' @export
+condDirichlet <- function(){
+  cat("condDirichlet() has been removed\n")
+  cat("use elicitDirichlet() instead for eliciting a Dirichlet distribution\n")
+  cat("and exploring conditional distribuitions.\n")
+}
+
+#' @export
+elicitConcProb <- function(){
+  cat("elicitConcProb() has been removed\n")
+  cat("Use elicitBivariate() instead. \n")  
+}