+++ R/elicitConcProb.R
@@ -1,5 +1,0 @@
-#' @export
-elicitConcProb <- function(){
-  cat("elicitConcProb() has been removed\n")
-  cat("Use elicitBivariate() instead. \n")  
-}