+++ R/elicitQuartiles.R
@@ -1,6 +1,0 @@
-
-#' @export
-elicitQuartiles <- function(){
-  cat("elicitQuartiles() has been removed\n")
-  cat("quartiles elicitation can be performed using elicit()\n")  
-  }