+++ R/elicitTertiles.R
@@ -1,5 +1,0 @@
-#' @export
-elicitTertiles <- function(){
-  cat("elicitTertiles() has been removed\n")
-  cat("tertiles elicitation can be performed using elicit()\n")  
-}