+++ NEWS.md
@@ -1,4 +1,4 @@
-SHELF v1.9.0.9000 (2024-02-27) 
+SHELF v1.10.0 (2024-05-07) 
 ==============================
 
 * Added the skew normal distribution to the set of fitted distributions.
@@ -16,6 +16,8 @@ SHELF v1.9.0.9000 (2024-02-27)
 * package test coverage improved.
 
 * elicitation report includes roulette allocation, if used.
+
+* improved error handling in shiny apps.
 
 SHELF v1.9.0 (2023-05-31) 
 ==============================